pyproject.toml
[tool.poetry]
name = "listening-monster"
version = "0.1.0"
description = "An audio trancription web app"
authors = ["Leo Ngari <ngarimwenda@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
audiosegment = "^0.23.0"
flask = "^3.0.0"
speechrecognition = "^3.10.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
